+++ lib/src/analyzers/lint_analyzer/lint_analyzer.dart
@@ -105,7 +105,7 @@ class LintAnalyzer {
           filePaths.intersection(context.contextRoot.analyzedFiles().toSet());
 
       for (final filePath in analyzedFiles) {
-        final unit = await context.currentSession.getResolvedUnit(filePath);
+        final unit = await context.currentSession.getResolvedUnit2(filePath);
         if (unit is ResolvedUnitResult) {
           final result = _runAnalysisForFile(
             unit,
@@ -130,14 +130,19 @@ class LintAnalyzer {
     String rootFolder, {
     String? filePath,
   }) {
-    if (result.state == ResultState.VALID &&
+    final unit = result.unit;
+    final content = result.content;
+
+    if (unit != null &&
+        content != null &&
+        result.state == ResultState.VALID &&
         filePath != null &&
         _isSupported(result)) {
-      final ignores = Suppression(result.content, result.lineInfo);
+      final ignores = Suppression(content, result.lineInfo);
       final internalResult = InternalResolvedUnitResult(
         filePath,
-        result.content,
-        result.unit,
+        content,
+        unit,
         result.lineInfo,
       );
       final relativePath = relative(filePath, from: rootFolder);
@@ -433,5 +438,7 @@ class LintAnalyzer {
   }
 
   bool _isSupported(AnalysisResult result) =>
-      result.path.endsWith('.dart') && !result.path.endsWith('.g.dart');
+      result.path != null &&
+      result.path!.endsWith('.dart') &&
+      !result.path!.endsWith('.g.dart');
 }