+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 4.2.0-dev.4
+
+* Revert: Changed the supported `analyzer` version to `^2.1.0`.
+
 ## 4.2.0-dev.3
 
 * Changed the supported `analyzer` version to `^2.1.0`.